Add windows platform
