+++ wiretunn/Cargo.toml
@@ -15,6 +15,7 @@ libc = "0.2.153"
 serde = { version = "1.0.197", features = ["derive"] }
 serde_with = "3.7.0"
 toml = "0.8.12"
+byteorder = "1.5.0"
 base64 = "0.22.0"
 bytes = "1.6.0"
 num_cpus = { version = "1", optional = true }
