+++ wiretunn/src/tun/mod.rs
@@ -2,6 +2,7 @@
 use std::os::unix::io::RawFd;
 
 use ipnet::IpNet;
+use std::net::IpAddr;
 use tokio::{
     io::{AsyncReadExt, AsyncWriteExt},
     sync::mpsc,
@@ -23,6 +24,11 @@ pub struct TunBuilder {
 impl TunBuilder {
     pub fn tun_name(&mut self, name: &str) {
         self.tun_config.tun_name(name);
+
+        #[cfg(target_os = "windows")]
+        self.tun_config.platform_config(|config| {
+            config.device_guid(self::sys::deterministic_guid(name));
+        });
     }
 
     pub fn address(&mut self, addr: IpNet) {
@@ -51,11 +57,6 @@ impl TunBuilder {
             config.ensure_root_privileges(true);
         });
 
-        #[cfg(target_os = "windows")]
-        config.platform_config(|config| {
-            config.device_guid(deterministic_guid(config.name()));
-        });
-
         let device = tun2::create_as_async(&self.tun_config)?;
 
         Ok(Tun { device })
@@ -69,6 +70,10 @@ pub struct Tun {
 impl Tun {
     pub fn tun_name(&self) -> Result<String, Error> {
         self.device.as_ref().tun_name().map_err(Error::TunError)
+    }
+
+    pub fn address(&self) -> Result<IpAddr, Error> {
+        self.device.as_ref().address().map_err(Error::TunError)
     }
 
     #[allow(clippy::type_complexity)]
@@ -115,25 +120,4 @@ impl Tun {
 
         Ok((iface_input, iface_output))
     }
-}
-
-#[cfg(windows)]
-fn deterministic_guid(ifname: &str) -> Option<u128> {
-    use blake2::{Blake2s256, Digest};
-    use byteorder::{ByteOrder, LittleEndian};
-    use bytes::{BufMut, BytesMut};
-    use unicode_normalization::UnicodeNormalization;
-
-    // https://github.com/WireGuard/wireguard-windows/blob/master/tunnel/deterministicguid.go
-    let mut data = BytesMut::from("Fixed WireGuard Windows GUID v1 jason@zx2c4.com");
-    let b2str = ifname.nfc().map(|c| c as u8).collect::<Vec<u8>>();
-    let mut b2num = [0u8; 4];
-    LittleEndian::write_u32(&mut b2num, b2str.len() as u32);
-    data.put_slice(&b2num[..]);
-    data.put_slice(&b2str[..]);
-
-    let mut buf = [0u8; 16];
-    buf.copy_from_slice(&Blake2s256::digest(&data[..])[..16]);
-
-    Some(u128::from_le_bytes(buf))
 }
